Cleanup deployment for red-bank and rover.
