+++ scripts/deploy/base/index-only-red-bank.ts
@@ -1,7 +1,6 @@
 import { setupDeployer } from './setup-deployer'
-import { printGreen, printRed, printYellow } from '../../utils/chalk'
 import { DeploymentConfig } from '../../types/NEW_config'
-import { wasmFile } from '../../utils/environment'
+import { printGreen, printRed } from '../../utils/chalk'
 
 export interface TaskRunnerProps {
   config: DeploymentConfig
@@ -12,6 +11,7 @@ export const taskRunner = async ({ config, label }: TaskRunnerProps) => {
   const deployer = await setupDeployer(config, label)
 
   try {
+    await deployer.saveStorage()
     await deployer.assertDeployerBalance()
 
     // Upload contracts
@@ -25,13 +25,9 @@ export const taskRunner = async ({ config, label }: TaskRunnerProps) => {
     )
     await deployer.upload('swapper', `mars_swapper_${config.swapper.name}.wasm`)
     await deployer.upload('params', `mars_params.wasm`)
-    await deployer.upload('accountNft', wasmFile('mars_account_nft'))
-    await deployer.upload('mockVault', wasmFile('mars_mock_vault'))
-    await deployer.upload('zapper', wasmFile(config.zapperContractName))
-    await deployer.upload('creditManager', wasmFile('mars_credit_manager'))
-    await deployer.upload('health', wasmFile('mars_rover_health'))
 
     // Instantiate contracts
+    deployer.setOwnerAddr() // TODO can be removed, see index.ts
     await deployer.instantiateAddressProvider()
     await deployer.instantiateRedBank()
     await deployer.instantiateIncentives()
@@ -39,15 +35,6 @@ export const taskRunner = async ({ config, label }: TaskRunnerProps) => {
     await deployer.instantiateRewards()
     await deployer.instantiateSwapper()
     await deployer.instantiateParams()
-    await deployer.instantiateMockVault()
-    await deployer.instantiateZapper()
-    await deployer.instantiateHealthContract()
-    await deployer.instantiateCreditManager()
-    await deployer.instantiateNftContract()
-    await deployer.setConfigOnHealthContract()
-    await deployer.transferNftContractOwnership()
-    await deployer.setConfigOnCreditManagerContract()
-    await deployer.updateAddressProviderWithNewAddrs()
     await deployer.saveDeploymentAddrsToFile(label)
 
     // setup
@@ -64,40 +51,15 @@ export const taskRunner = async ({ config, label }: TaskRunnerProps) => {
       await deployer.setOracle(oracleConfig)
     }
 
-    await deployer.grantCreditLines()
-
-    // Test basic user flows
-    if (config.runTests && config.testActions) {
+    // run tests
+    if (config.runTests) {
       await deployer.executeDeposit()
       await deployer.executeBorrow()
       await deployer.executeRepay()
       await deployer.executeWithdraw()
       // await deployer.executeRewardsSwap()
-
-      const rover = await deployer.newUserRoverClient(config.testActions)
-      await rover.createCreditAccount()
-      await rover.deposit()
-      await rover.lend()
-      await rover.borrow()
-      await rover.swap()
-      await rover.repay()
-      await rover.reclaim()
-      await rover.withdraw()
-
-      const vaultConfig = config.vaults[0].vault
-      const info = await rover.getVaultInfo(vaultConfig)
-      await rover.zap(info.tokens.base_token)
-      await rover.vaultDeposit(vaultConfig, info)
-      if (info.lockup) {
-        await rover.vaultRequestUnlock(vaultConfig, info)
-      } else {
-        await rover.vaultWithdraw(vaultConfig, info)
-        await rover.unzap(info.tokens.base_token)
-      }
-      await rover.refundAllBalances()
     }
 
-    // If multisig is set, transfer ownership from deployer to multisig
     if (config.multisigAddr) {
       await deployer.updateIncentivesContractOwner()
       await deployer.updateRedBankContractOwner()
@@ -106,12 +68,10 @@ export const taskRunner = async ({ config, label }: TaskRunnerProps) => {
       await deployer.updateSwapperContractOwner()
       await deployer.updateParamsContractOwner()
       await deployer.updateAddressProviderContractOwner()
-      await deployer.updateCreditManagerOwner()
-      await deployer.updateHealthOwner()
       printGreen('It is confirmed that all contracts have transferred ownership to the Multisig')
+    } else {
+      printGreen('Owner remains the deployer address.')
     }
-
-    printYellow('COMPLETE')
   } catch (e) {
     printRed(e)
   } finally {
