+++ scripts/deploy/osmosis/testnet-index.ts
@@ -2,5 +2,8 @@ import { taskRunner } from '../base'
 import { osmosisTestnetConfig } from './testnet-config'
 
 void (async function () {
-  await taskRunner(osmosisTestnetConfig)
+  await taskRunner({
+    config: osmosisTestnetConfig,
+    label: 'deployer-owner',
+  })
 })()
